move exit node ping to module
